Cambiando hbm2ddl a update, cambiando id de domainentity a integer, y agregando algunos algo base de la red.

+++ trunk/snanalizer/java-src/snanalizer/data/BaseRepository.java
@@ -8,10 +8,10 @@ public interface BaseRepository<E extends DomainEntity> {
 
 	public List<E> getAll();
 
-	public E getById(Long id);
+	public E getById(Integer id);
 
 	public void add(E newEntity);
 
-	public void remove(Long id);
+	public void remove(Integer id);
 
 }
